3619: Rename SmartDefaultAttributeEditor files
