+++ common/src/View/SmartDefaultPropertyEditor.h
